iva propios en lentes y articulos resueltos

+++ DisplaBackend/DisplaBackend/Models/CategoriaIva.cs
@@ -12,8 +12,8 @@ namespace DisplaBackend.Models
 
         public int Id { get; set; }
         public string Descripcion { get; set; }
-        public double? Tasa { get; set; }
-        public double? SobreTasa { get; set; }
+        public decimal? Tasa { get; set; }
+        public decimal? SobreTasa { get; set; }
         public bool? Discrimina { get; set; }
         public bool? Recateg { get; set; }
         public int? CodigoRece { get; set; }
